Increase the number of getContexts
EndGetContext is not the end of the native operation, its the end of the
managed one.
When the native operation completes it signals HttpListener, which did
postprocessing and then invoked your callback. 10 threads may not be
enough due to the postprocessing lag her and hence we use a processor
multiplier.